feat : API 구현

+++ src/main/java/com/example/simple_board/common/response/GetBoardListResponse.java
@@ -6,13 +6,13 @@ import java.util.List;
 
 public record GetBoardListResponse(
         @Schema(description = "요청한 페이지 index", example = "1")
-        int page,
+        long page,
         @Schema(description = "조회된 게시글 목록 크기", example = "10")
         int size,
         @Schema(description = "총 게시글 수", example = "1000")
-        int totalElement,
+        long totalElement,
         @Schema(description = "총 게시글 페이지 수", example = "100")
-        int totalSize,
+        long  totalSize,
         @Schema(description = "조회된 게시글 id 정보 목록")
         List<Long> boardIdList
 ) {
